+++ resources/js/components/ContactCard.vue
@@ -1,19 +1,21 @@
-   <template>
-   <div class="card p-3" style="max-width: 540px;">
-      <div class="row no-gutters">
-         <div class="col-md-3 col-lg-3">
-            <img :src="contact.image" class="rounded my-auto" width="85" height="110"/>
-         </div>
-         <div class="col-md-6">
-            <div class="card-body">
-               <h4 class="card-title">{{ contact.name }}</h4>
-               <p class="card-text mb-0">{{ contact.number }}</p>
-               <p class="card-text">{{ contact.email }}</p>
+<template>
+   <div>
+      <div class="card p-3" style="max-width: 540px;">
+         <div class="row no-gutters">
+            <div class="col-md-3 col-lg-3">
+               <img :src="contact.image" class="rounded my-auto" width="85" height="110"/>
             </div>
-         </div>
-         <div class="col-md-3">
-            <button class="btn btn-outline-primary mt-1 mb-3"><i class="fa fa-pencil"></i></button>
-            <button @click="deleteContact(contact.id)" class="btn btn-danger"><i class="fa fa-trash"></i></button>
+            <div class="col-md-6">
+               <div class="card-body">
+                  <h4 class="card-title">{{ contact.name }}</h4>
+                  <p class="card-text mb-0">{{ contact.number }}</p>
+                  <p class="card-text">{{ contact.email }}</p>
+               </div>
+            </div>
+            <div class="col-md-3">
+               <button class="btn btn-outline-primary mt-1 mb-3"><i class="fa fa-pencil"></i></button>
+               <button @click="deleteContact(contact.id)" class="btn btn-danger"><i class="fa fa-trash"></i></button>
+            </div>
          </div>
       </div>
    </div>
@@ -26,14 +28,17 @@ export default {
    },
    methods: {
       deleteContact(id) {
-         let contacts = JSON.parse(localStorage.getItem('contacts'))
-         let index = contacts.findIndex(c => c.id === id)
-         
-         contacts.splice(index, 1)
-         localStorage.setItem('contacts', JSON.stringify(contacts))
-         
-         this.$root.$emit('contactDeleted')
-      }
+         let result = confirm(`Do you want to delete ${this.contact.name}'s contact?`)
+         if(result) {
+            let contacts = JSON.parse(localStorage.getItem('contacts'))
+            let index = contacts.findIndex(c => c.id === id)
+            
+            contacts.splice(index, 1)
+            localStorage.setItem('contacts', JSON.stringify(contacts))
+
+            this.$root.$emit('contactDeleted')
+         }
+      },
    },
 }
 </script>
